pyproject.toml
[tool.poetry]
name = "sre_archtest"
version = "0.1.0"
description = ""
authors = ["Philippe Gonzaga <francisco.philippe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
kafka-python = "^2.0.2"
python-decouple = "^3.4"
uvicorn = "^0.13.4"
motor = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
ipython = "^7.22.0"
pytest-async-mongodb = "^0.0.1-alpha.3"
requests = "^2.25.1"
pytest-asyncio = "^0.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
